+++ changelog.md
@@ -1,15 +1,22 @@
 ### v0.7.0
 
 ##### Breaking Changes
- None.
+
+* The network hierarchy has been updated to use CIM classes.
+* `getFeeder` has been deprecated.
 
 ##### New Features
+
 * Added normal and current version of the connected equipment trace. See `Tracing` for details.
+* Added a generic API call for populating the `Equipment` of an `EquipmentContainer`.
+* Added API calls for getting loops.
 
 ##### Enhancements
-* None.
+
+* The network hierarchy now contains circuits and loops.
 
 ##### Fixes
+
 * None.
 
 ##### Notes